+++ jkube-kit/config/service/src/test/java/org/eclipse/jkube/kit/config/service/openshift/OpenshiftBuildServiceIntegrationTest.java
@@ -28,7 +28,9 @@ import io.fabric8.openshift.api.model.ImageStream;
 import io.fabric8.openshift.api.model.ImageStreamBuilder;
 import io.fabric8.openshift.api.model.ImageStreamStatusBuilder;
 import io.fabric8.openshift.api.model.NamedTagEventListBuilder;
-import io.fabric8.openshift.client.server.mock.OpenShiftServer;
+import io.fabric8.openshift.client.NamespacedOpenShiftClient;
+import io.fabric8.openshift.client.server.mock.OpenShiftMockServer;
+import okhttp3.mockwebserver.RecordedRequest;
 import org.apache.commons.io.FileUtils;
 import org.eclipse.jkube.kit.build.api.assembly.JKubeBuildTarArchiver;
 import org.eclipse.jkube.kit.common.RegistryConfig;
@@ -45,16 +47,13 @@ import org.eclipse.jkube.kit.config.service.BuildServiceConfig;
 import org.eclipse.jkube.kit.config.service.JKubeServiceException;
 import mockit.Expectations;
 import mockit.Mocked;
-import okhttp3.mockwebserver.MockWebServer;
-import okhttp3.mockwebserver.RecordedRequest;
 
 import org.eclipse.jkube.kit.config.service.JKubeServiceHub;
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.IOException;
@@ -70,12 +69,11 @@ import static org.junit.Assert.assertTrue;
 
 public class OpenshiftBuildServiceIntegrationTest {
 
-    private static final Logger LOG = LoggerFactory.getLogger(OpenshiftBuildServiceIntegrationTest.class);
-
     private static final int MAX_TIMEOUT_RETRIES = 5;
 
-    @Rule
-    public final OpenShiftServer mockServer = new OpenShiftServer(false);
+    private OpenShiftMockServer mockServer;
+
+    private NamespacedOpenShiftClient openShiftClient;
 
     @Rule
     public TemporaryFolder temporaryFolder = new TemporaryFolder();
@@ -116,13 +114,15 @@ public class OpenshiftBuildServiceIntegrationTest {
         FileUtils.deleteDirectory(new File(baseDir, projectName));
         final File dockerFile = new File(baseDir, "Docker.tar");
         FileUtils.touch(dockerFile);
+        mockServer = new OpenShiftMockServer(false);
+        openShiftClient = mockServer.createOpenShiftClient();
 
         // @formatter:off
         new Expectations(openshiftHelper) {{
             // Partial Mock
-            openshiftHelper.isOpenShift(mockServer.getOpenshiftClient()); result = true;
+            openshiftHelper.isOpenShift(openShiftClient); result = true;
 
-            clusterAccess.createDefaultClient(); result = mockServer.getOpenshiftClient();
+            clusterAccess.createDefaultClient(); result = openShiftClient;
             jKubeServiceHub.getLog(); result = logger;
             jKubeServiceHub.getDockerServiceHub(); minTimes = 0;
             jKubeServiceHub.getDockerServiceHub().getArchiveService(); minTimes = 0;
@@ -168,6 +168,12 @@ public class OpenshiftBuildServiceIntegrationTest {
                 .buildOutputKind("DockerImage");
     }
 
+    @After
+    public void cleanUp() {
+        openShiftClient.close();
+        mockServer.destroy();
+    }
+
     @Test
     public void testSuccessfulBuild() throws Exception {
         retryInMockServer(() -> {
@@ -177,7 +183,7 @@ public class OpenshiftBuildServiceIntegrationTest {
             new OpenshiftBuildService(jKubeServiceHub).build(image);
 
             // we should find a better way to assert that a certain call has been made
-            assertTrue(mockServer.getMockServer().getRequestCount() > 8);
+            assertTrue(mockServer.getRequestCount() > 8);
             collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
             assertEquals("{\"apiVersion\":\"build.openshift.io/v1\",\"kind\":\"BuildConfig\",\"metadata\":{\"name\":\"myapp-s2i-suffix2\"},\"spec\":{\"output\":{\"to\":{\"kind\":\"ImageStreamTag\",\"name\":\"myapp:latest\"}},\"source\":{\"type\":\"Binary\"},\"strategy\":{\"sourceStrategy\":{\"forcePull\":false,\"from\":{\"kind\":\"DockerImage\",\"name\":\"myapp\"}},\"type\":\"Source\"}}}", collector.getBodies().get(1));
             collector.assertEventsNotRecorded("patch-build-config");
@@ -195,7 +201,7 @@ public class OpenshiftBuildServiceIntegrationTest {
             new OpenshiftBuildService(jKubeServiceHub).build(image);
 
             // we should find a better way to assert that a certain call has been made
-            assertTrue(mockServer.getMockServer().getRequestCount() > 8);
+            assertTrue(mockServer.getRequestCount() > 8);
             collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
             assertEquals("{\"apiVersion\":\"build.openshift.io/v1\",\"kind\":\"BuildConfig\",\"metadata\":{\"name\":\"myapp-s2i\"},\"spec\":{\"output\":{\"to\":{\"kind\":\"ImageStreamTag\",\"name\":\"myapp:latest\"}},\"source\":{\"type\":\"Binary\"},\"strategy\":{\"sourceStrategy\":{\"forcePull\":false,\"from\":{\"kind\":\"DockerImage\",\"name\":\"myapp\"}},\"type\":\"Source\"}}}", collector.getBodies().get(1));
             collector.assertEventsNotRecorded("patch-build-config");
@@ -215,7 +221,7 @@ public class OpenshiftBuildServiceIntegrationTest {
 
             new OpenshiftBuildService(jKubeServiceHub).build(image);
 
-            assertTrue(mockServer.getMockServer().getRequestCount() > 8);
+            assertTrue(mockServer.getRequestCount() > 8);
             collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
             assertEquals("{\"apiVersion\":\"build.openshift.io/v1\",\"kind\":\"BuildConfig\",\"metadata\":{\"name\":\"myapp-docker\"},\"spec\":{\"output\":{\"to\":{\"kind\":\"ImageStreamTag\",\"name\":\"myapp:latest\"}},\"source\":{\"type\":\"Binary\"},\"strategy\":{\"dockerStrategy\":{\"from\":{\"kind\":\"DockerImage\",\"name\":\"myapp\"},\"noCache\":false},\"type\":\"Docker\"}}}", collector.getBodies().get(1));
             collector.assertEventsNotRecorded("patch-build-config");
@@ -236,7 +242,7 @@ public class OpenshiftBuildServiceIntegrationTest {
 
             new OpenshiftBuildService(jKubeServiceHub).build(image);
 
-            assertTrue(mockServer.getMockServer().getRequestCount() > 8);
+            assertTrue(mockServer.getRequestCount() > 8);
             collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
             assertEquals("{\"apiVersion\":\"build.openshift.io/v1\",\"kind\":\"BuildConfig\",\"metadata\":{\"name\":\"component1-component2-name-docker\"},\"spec\":{\"output\":{\"to\":{\"kind\":\"ImageStreamTag\",\"name\":\"component1-component2-name:tag\"}},\"source\":{\"type\":\"Binary\"},\"strategy\":{\"dockerStrategy\":{\"from\":{\"kind\":\"DockerImage\",\"name\":\"myapp\"},\"noCache\":false},\"type\":\"Docker\"}}}", collector.getBodies().get(1));
             collector.assertEventsNotRecorded("patch-build-config");
@@ -256,7 +262,7 @@ public class OpenshiftBuildServiceIntegrationTest {
 
             new OpenshiftBuildService(jKubeServiceHub).build(image);
 
-            assertTrue(mockServer.getMockServer().getRequestCount() > 8);
+            assertTrue(mockServer.getRequestCount() > 8);
             collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
             assertEquals("{\"apiVersion\":\"build.openshift.io/v1\",\"kind\":\"BuildConfig\",\"metadata\":{\"name\":\"myapp\"},\"spec\":{\"output\":{\"to\":{\"kind\":\"ImageStreamTag\",\"name\":\"myapp:latest\"}},\"source\":{\"type\":\"Binary\"},\"strategy\":{\"dockerStrategy\":{\"from\":{\"kind\":\"DockerImage\",\"name\":\"myapp\"},\"noCache\":false},\"type\":\"Docker\"}}}", collector.getBodies().get(1));
             collector.assertEventsNotRecorded("patch-build-config");
@@ -289,7 +295,7 @@ public class OpenshiftBuildServiceIntegrationTest {
 
             service.build(fromExtImage);
 
-            assertTrue(mockServer.getMockServer().getRequestCount() > 8);
+            assertTrue(mockServer.getRequestCount() > 8);
             collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
             assertEquals("{\"apiVersion\":\"build.openshift.io/v1\",\"kind\":\"BuildConfig\",\"metadata\":{\"name\":\"myapp-docker\"},\"spec\":{\"output\":{\"to\":{\"kind\":\"ImageStreamTag\",\"name\":\"myapp:latest\"}},\"source\":{\"type\":\"Binary\"},\"strategy\":{\"dockerStrategy\":{\"from\":{\"kind\":\"ImageStreamTag\",\"name\":\"app:1.2-1\",\"namespace\":\"my-project\"},\"noCache\":true},\"type\":\"Docker\"}}}", collector.getBodies().get(1));
             collector.assertEventsNotRecorded("patch-build-config");
@@ -305,7 +311,7 @@ public class OpenshiftBuildServiceIntegrationTest {
             new OpenshiftBuildService(jKubeServiceHub).build(image);
 
             // we should find a better way to assert that a certain call has been made
-            assertTrue(mockServer.getMockServer().getRequestCount() > 8);
+            assertTrue(mockServer.getRequestCount() > 8);
             collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
             collector.assertEventsNotRecorded("patch-build-config");
         });
@@ -377,10 +383,10 @@ public class OpenshiftBuildServiceIntegrationTest {
             new OpenshiftBuildService(jKubeServiceHub).build(image);
 
             // we should add a better way to assert that a certain call has been made
-            assertTrue(mockServer.getMockServer().getRequestCount() > 8);
+            assertTrue(mockServer.getRequestCount() > 7);
             collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
             assertEquals("{\"apiVersion\":\"build.openshift.io/v1\",\"kind\":\"BuildConfig\",\"metadata\":{\"name\":\"myapp-s2i-suffix2\"},\"spec\":{\"output\":{\"to\":{\"kind\":\"DockerImage\",\"name\":\"myapp:latest\"}},\"source\":{\"type\":\"Binary\"},\"strategy\":{\"sourceStrategy\":{\"forcePull\":false,\"from\":{\"kind\":\"DockerImage\",\"name\":\"myapp\"}},\"type\":\"Source\"}}}", collector.getBodies().get(1));
-                        collector.assertEventsNotRecorded("patch-build-config");
+            collector.assertEventsNotRecorded("patch-build-config");
             assertFalse(containsRequest("imagestreams"));
         });
     }
@@ -394,7 +400,7 @@ public class OpenshiftBuildServiceIntegrationTest {
             new OpenshiftBuildService(jKubeServiceHub).build(image);
 
             // we should find a better way to assert that a certain call has been made
-            assertTrue(mockServer.getMockServer().getRequestCount() > 8);
+            assertTrue(mockServer.getRequestCount() > 7);
             collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
             assertEquals("{\"apiVersion\":\"build.openshift.io/v1\",\"kind\":\"BuildConfig\",\"metadata\":{\"name\":\"myapp-s2i-suffix2\"},\"spec\":{\"output\":{\"pushSecret\":{\"name\":\"pushsecret-fabric8\"},\"to\":{\"kind\":\"DockerImage\",\"name\":\"myapp:latest\"}},\"source\":{\"type\":\"Binary\"},\"strategy\":{\"sourceStrategy\":{\"forcePull\":false,\"from\":{\"kind\":\"DockerImage\",\"name\":\"myapp\"}},\"type\":\"Source\"}}}", collector.getBodies().get(1));
             collector.assertEventsNotRecorded("patch-build-config");
@@ -577,12 +583,9 @@ public class OpenshiftBuildServiceIntegrationTest {
         return result;
     }
 
-    /**
-     * Helper method to verify request url path
-     */
     private Boolean containsRequest(String path)   {
         try {
-            MockWebServer mock = mockServer.getMockServer();
+            OpenShiftMockServer mock = mockServer;
             int count = mock.getRequestCount();
             RecordedRequest request = null;
             while ( count-- > 0 ) {
@@ -596,5 +599,4 @@ public class OpenshiftBuildServiceIntegrationTest {
         }
         return false;
     }
-
 }