+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Usage:
 ```
 ### 1.6.0-SNAPSHOT
 * Fix #887: Incorrect warning about overriding environment variable
+* Fix #802: Update Fabric8 kubernetes Client to v5.10.1
 
 ### 1.5.1 (2021-10-28)
 * Fix #1084: Gradle dependencies should be test or provided scope